api/requirements.txt
amqp==5.0.6
aniso8601==7.0.0
argon2-cffi==20.1.0
asgiref==3.3.4
asyncio==3.4.3
billiard==3.6.4.0
celery==5.0.5
cffi==1.14.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
dnspython==2.0.0
email-validator==1.1.2
fastapi==0.63.0
fastapi-jwt-auth==0.5.0
graphene==2.1.8
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.12.0
idna==3.0
implicit==0.4.4
joblib==1.0.1
kombu==5.0.2
motor==2.3.0
numpy==1.20.2
odmantic==0.3.2
pandas==1.2.4
passlib==1.7.4
promise==2.3
prompt-toolkit==3.0.18
pycparser==2.20
pydantic==1.7.3
PyJWT==1.7.1
pymongo==3.11.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-multipart==0.0.5
pytz==2021.1
redis==3.5.3
Rx==1.6.1
scikit-learn==0.24.2
scipy==1.6.2
six==1.15.0
starlette==0.13.6
threadpoolctl==2.1.0
tqdm==4.60.0
uvicorn==0.13.3
vine==5.0.0
wcwidth==0.2.5
